fix(davinci): unblock CI and share the S2 slot test harness
Fix clippy::manual_contains / clippy::op_ref, replace emit_comp
contains() pins with exact oracles, and split emit modules plus the
atelier_dom differential harness so the 350-line ratchet stays green.
