+++ crates/vize_atelier_core/src/codegen/element/block.rs
@@ -19,8 +19,7 @@ use super::{
         props::generate_props,
         slots::{
             generate_slot_outlet_name, generate_slot_outlet_props, generate_slots,
-            has_dynamic_slots_flag, has_forwarded_slot_outlet, has_slot_children,
-            has_slot_outlet_props,
+            has_slot_children, has_slot_outlet_props, needs_dynamic_slots_patch,
         },
     },
     directives::{
@@ -340,12 +339,7 @@ pub fn generate_element_block(ctx: &mut CodegenContext, el: &ElementNode<'_>) {
 
             // Add DYNAMIC_SLOTS flag (1024) if component has dynamic slots
             // KeepAlive always gets DYNAMIC_SLOTS
-            if el.tag == "KeepAlive"
-                || el.tag == "keep-alive"
-                || (ctx.in_v_for && has_slot_children(el))
-                || has_dynamic_slots_flag(el, &ctx.source)
-                || (ctx.has_slot_params() && has_forwarded_slot_outlet(el))
-            {
+            if needs_dynamic_slots_patch(ctx, el) {
                 let dynamic_slots_flag = 1024;
                 patch_flag = Some(patch_flag.unwrap_or(0) | dynamic_slots_flag);
             }
