+++ crates/vize_atelier_core/src/codegen/slots/name.rs
@@ -1,8 +1,21 @@
-use crate::{DirectiveNode, ExpressionNode};
+use crate::{DirectiveNode, ElementNode, ExpressionNode, PropNode};
 
 use super::super::context::CodegenContext;
 use super::super::expression::generate_expression;
-use super::super::helpers::escape_js_string;
+use super::super::helpers::{escape_js_string, is_valid_js_identifier};
+
+pub(super) fn component_root_slot<'a, 'b>(
+    el: &'b ElementNode<'a>,
+) -> Option<&'b DirectiveNode<'a>> {
+    el.props.iter().find_map(|p| {
+        if let PropNode::Directive(dir) = p
+            && dir.name == "slot"
+        {
+            return Some(dir.as_ref());
+        }
+        None
+    })
+}
 
 pub(super) fn generate_slot_entry_name(
     ctx: &mut CodegenContext,
@@ -17,5 +30,26 @@ pub(super) fn generate_slot_entry_name(
             ctx.push(&escape_js_string(slot_name));
             ctx.push("\"");
         }
+    }
+}
+
+pub(super) fn emit_slot_property_name(
+    ctx: &mut CodegenContext,
+    slot_dir: &DirectiveNode<'_>,
+    slot_name: &str,
+    is_dynamic: bool,
+) {
+    if is_dynamic {
+        ctx.push("[");
+        if let Some(arg) = &slot_dir.arg {
+            generate_expression(ctx, arg);
+        }
+        ctx.push("]");
+    } else if is_valid_js_identifier(slot_name) {
+        ctx.push(slot_name);
+    } else {
+        ctx.push("\"");
+        ctx.push(&escape_js_string(slot_name));
+        ctx.push("\"");
     }
 }
