+++ crates/vize_ricalco/src/emit/props.rs
@@ -2,18 +2,15 @@
 
 use alloc::vec::Vec as StdVec;
 
-use vize_carton::{Span, String};
-use vize_disegno::expr::{ExprRef, JsExpr};
-use vize_disegno::op::{Attribute, BindOp, BindingOp, DynamicName, OnOp};
+use vize_carton::String;
+use vize_disegno::op::{Attribute, BindingOp};
 
 use super::EmitCx;
 use super::EmitError;
-use super::buf::Buf;
-use super::js::{escape_js_string, push_ident_key};
-use super::on::{
-    admit_on, emit_on_pair, emit_on_value, event_key_for, is_inline_handler_source,
-    needs_hydration, wraps_on,
-};
+use super::on::{admit_on, event_key_for, needs_hydration};
+
+pub(super) use super::props_bind::{js_value, static_bind_name};
+pub(super) use super::props_object::{Piece, emit_props_object, pieces};
 
 pub(super) struct Patch {
     pub flag: i32,
@@ -39,7 +36,9 @@ pub(super) fn admit_bindings(
                 if name == "ref" || !bind.modifiers.is_empty() {
                     return Err(EmitError::Unsupported);
                 }
-                let ExprRef::Js(_) = bind.value.ok_or(EmitError::Unsupported)? else {
+                let vize_disegno::expr::ExprRef::Js(_) =
+                    bind.value.ok_or(EmitError::Unsupported)?
+                else {
                     return Err(EmitError::Unsupported);
                 };
                 match name {
@@ -147,349 +146,6 @@ pub(super) fn emit_bind_props(
     )
 }
 
-pub(super) fn emit_props_object(
-    cx: &mut EmitCx<'_>,
-    pieces: &[Piece<'_>],
-    if_key: Option<&str>,
-    skip_normalize: bool,
-    empty_key_multiline: bool,
-    is_plain_element: bool,
-) -> Result<(), EmitError> {
-    let skip_class = pieces_have_named(pieces, "class");
-    let skip_key = if_key.is_some();
-    let visible: StdVec<&Piece<'_>> = pieces
-        .iter()
-        .filter(|piece| !skip_emitted_key(piece, if_key, skip_class, skip_key))
-        .collect();
-    if let Some(key) = if_key
-        && visible.is_empty()
-    {
-        if empty_key_multiline {
-            cx.buf.push("{");
-            cx.buf.indent();
-            cx.buf.newline();
-            cx.buf.push("key: ");
-            cx.buf.push(key);
-            cx.buf.deindent();
-            cx.buf.newline();
-            cx.buf.push("}");
-        } else {
-            cx.buf.push("{ key: ");
-            cx.buf.push(key);
-            cx.buf.push(" }");
-        }
-        return Ok(());
-    }
-    let extra = usize::from(if_key.is_some());
-    let multiline = visible.len() + extra > 1
-        || pieces_have_named(pieces, "class")
-        || pieces_have_named(pieces, "style")
-        || pieces_have_inline_on(pieces);
-    if multiline {
-        cx.buf.push("{");
-        cx.buf.indent();
-    } else {
-        cx.buf.push("{ ");
-    }
-    let mut i = 0;
-    if let Some(key) = if_key {
-        if multiline {
-            cx.buf.newline();
-        }
-        cx.buf.push("key: ");
-        cx.buf.push(key);
-        i = 1;
-    }
-    let mut emitted_merged = StdVec::new();
-    for piece in visible.iter() {
-        if let Some(key) = piece_event_key(piece, is_plain_element)
-            && merge_count(&visible, key.as_str(), is_plain_element) > 1
-        {
-            if emitted_merged.contains(&key) {
-                continue;
-            }
-            if i > 0 {
-                cx.buf.push(",");
-            }
-            if multiline {
-                cx.buf.newline();
-            } else if i > 0 {
-                cx.buf.push(" ");
-            }
-            emit_merged_handlers(cx, &visible, key.as_str(), is_plain_element)?;
-            emitted_merged.push(key);
-            i += 1;
-            continue;
-        }
-        if i > 0 {
-            cx.buf.push(",");
-        }
-        if multiline {
-            cx.buf.newline();
-        } else if i > 0 {
-            cx.buf.push(" ");
-        }
-        match piece {
-            Piece::Attr(attr) => emit_static_pair(cx, attr),
-            Piece::Bind(bind) => emit_bind_pair(cx, pieces, bind, skip_normalize)?,
-            Piece::On(on) => emit_on_pair(cx, on, is_plain_element)?,
-            Piece::ModelValue { name, source, .. } => super::model::emit_value(cx, name, source),
-            Piece::ModelUpdate { key, source, .. } => {
-                super::model::emit_update(cx, key.as_str(), source)
-            }
-            Piece::ModelModifiers {
-                name, modifiers, ..
-            } => super::model::emit_modifiers(cx, name.as_str(), modifiers),
-        }
-        i += 1;
-    }
-    if multiline {
-        cx.buf.deindent();
-        cx.buf.newline();
-        cx.buf.push("}");
-    } else {
-        cx.buf.push(" }");
-    }
-    Ok(())
-}
-
-pub(super) enum Piece<'a> {
-    Attr(&'a Attribute<'a>),
-    Bind(&'a BindOp<'a>),
-    On(&'a OnOp<'a>),
-    ModelValue {
-        name: &'a str,
-        source: &'a str,
-        span: Span,
-    },
-    ModelUpdate {
-        key: String,
-        source: &'a str,
-        span: Span,
-    },
-    ModelModifiers {
-        name: String,
-        modifiers: StdVec<&'a str>,
-        span: Span,
-    },
-}
-
-pub(super) fn pieces<'a>(
-    attributes: &'a [Attribute<'a>],
-    bindings: &'a [BindingOp<'a>],
-    skip_is: bool,
-) -> Result<StdVec<Piece<'a>>, EmitError> {
-    let mut out = StdVec::new();
-    for attr in attributes.iter() {
-        if skip_is && attr.name == "is" {
-            continue;
-        }
-        out.push(Piece::Attr(attr));
-    }
-    for binding in bindings.iter() {
-        match binding {
-            BindingOp::Bind(bind)
-                if skip_is && matches!(bind.name, Some(DynamicName::Static("is"))) => {}
-            BindingOp::Bind(bind) => out.push(Piece::Bind(bind)),
-            BindingOp::On(on) => out.push(Piece::On(on)),
-            BindingOp::Model(model) => super::model::expand(model, &mut out)?,
-            BindingOp::SlotContent(_) => {}
-            BindingOp::VueDirective(_) => {}
-            _ => return Err(EmitError::Unsupported),
-        }
-    }
-    out.sort_by_key(|piece| match piece {
-        Piece::Attr(attr) => attr.span.start,
-        Piece::Bind(bind) => bind.span.start,
-        Piece::On(on) => on.span.start,
-        Piece::ModelValue { span, .. }
-        | Piece::ModelUpdate { span, .. }
-        | Piece::ModelModifiers { span, .. } => span.start,
-    });
-    Ok(out)
-}
-
-pub(super) fn static_bind_name<'a>(bind: &'a BindOp<'a>) -> Result<&'a str, EmitError> {
-    match bind.name {
-        Some(DynamicName::Static(name)) => Ok(name),
-        Some(DynamicName::Dynamic(_)) | None => Err(EmitError::Unsupported),
-    }
-}
-
 fn has_attr(attributes: &[Attribute<'_>], name: &str) -> bool {
     attributes.iter().any(|attr| attr.name == name)
-}
-
-fn skip_emitted_key(
-    piece: &Piece<'_>,
-    if_key: Option<&str>,
-    skip_class: bool,
-    skip_key: bool,
-) -> bool {
-    match piece {
-        Piece::Attr(attr) => {
-            (skip_class && attr.name == "class") || (skip_key && attr.name == "key")
-        }
-        Piece::Bind(bind)
-            if skip_key
-                && matches!(bind.name, Some(DynamicName::Static("key")))
-                && js_value(bind).is_ok_and(|js| if_key == Some(js.source)) =>
-        {
-            true
-        }
-        _ => false,
-    }
-}
-
-fn pieces_have_named(pieces: &[Piece<'_>], name: &str) -> bool {
-    pieces.iter().any(|piece| match piece {
-        Piece::Bind(bind) => matches!(bind.name, Some(DynamicName::Static(n)) if n == name),
-        Piece::ModelValue { name: prop, .. } => *prop == name,
-        Piece::ModelModifiers { name: prop, .. } => prop.as_str() == name,
-        _ => false,
-    })
-}
-
-fn pieces_have_inline_on(pieces: &[Piece<'_>]) -> bool {
-    pieces.iter().any(|piece| match piece {
-        Piece::On(on) => {
-            wraps_on(on)
-                || match on.handler {
-                    Some(ExprRef::Js(js)) => is_inline_handler_source(js.source),
-                    _ => false,
-                }
-        }
-        Piece::ModelUpdate { .. } => true,
-        _ => false,
-    })
-}
-
-fn emit_static_pair(cx: &mut EmitCx<'_>, attr: &Attribute<'_>) {
-    push_ident_key(cx, attr.name);
-    cx.buf.push(": \"");
-    if let Some(value) = attr.value {
-        cx.buf.push(escape_js_string(value).as_str());
-    }
-    cx.buf.push("\"");
-}
-
-fn emit_bind_pair(
-    cx: &mut EmitCx<'_>,
-    pieces: &[Piece<'_>],
-    bind: &BindOp<'_>,
-    skip_normalize: bool,
-) -> Result<(), EmitError> {
-    let name = static_bind_name(bind)?;
-    let js = js_value(bind)?;
-    push_ident_key(cx, name);
-    cx.buf.push(": ");
-    match name {
-        "class" => emit_class_value(cx, pieces, bind, js, skip_normalize),
-        "style" => emit_style_value(cx, js, skip_normalize),
-        _ => cx.buf.push(js.source),
-    }
-    Ok(())
-}
-
-fn emit_class_value(
-    cx: &mut EmitCx<'_>,
-    pieces: &[Piece<'_>],
-    bind: &BindOp<'_>,
-    js: &JsExpr<'_>,
-    skip_normalize: bool,
-) {
-    if !skip_normalize {
-        cx.buf.use_normalize_class();
-        cx.buf.push(Buf::normalize_class_alias());
-        cx.buf.push("(");
-    }
-    if let Some(static_class) = pieces.iter().find_map(|piece| match piece {
-        Piece::Attr(attr) if attr.name == "class" => Some(*attr),
-        _ => None,
-    }) {
-        let before = static_class.span.start <= bind.span.start;
-        cx.buf.push("[");
-        if before {
-            cx.buf.push("\"");
-            cx.buf
-                .push(escape_js_string(static_class.value.unwrap_or("")).as_str());
-            cx.buf.push("\", ");
-            cx.buf.push(js.source);
-        } else {
-            cx.buf.push(js.source);
-            cx.buf.push(", \"");
-            cx.buf
-                .push(escape_js_string(static_class.value.unwrap_or("")).as_str());
-            cx.buf.push("\"");
-        }
-        cx.buf.push("]");
-    } else {
-        cx.buf.push(js.source);
-    }
-    if !skip_normalize {
-        cx.buf.push(")");
-    }
-}
-
-fn emit_style_value(cx: &mut EmitCx<'_>, js: &JsExpr<'_>, skip_normalize: bool) {
-    let object_literal = js.source.trim_start().starts_with('{');
-    let wrap = !skip_normalize && !object_literal;
-    if wrap {
-        cx.buf.use_normalize_style();
-        cx.buf.push(Buf::normalize_style_alias());
-        cx.buf.push("(");
-    }
-    cx.buf.push(js.source);
-    if wrap {
-        cx.buf.push(")");
-    }
-}
-
-fn piece_event_key(piece: &Piece<'_>, is_plain_element: bool) -> Option<String> {
-    match piece {
-        Piece::On(on) => event_key_for(on, is_plain_element).ok(),
-        Piece::ModelUpdate { key, .. } => Some(key.clone()),
-        _ => None,
-    }
-}
-
-fn merge_count(visible: &[&Piece<'_>], key: &str, is_plain_element: bool) -> usize {
-    visible
-        .iter()
-        .filter(|piece| piece_event_key(piece, is_plain_element).as_deref() == Some(key))
-        .count()
-}
-
-fn emit_merged_handlers(
-    cx: &mut EmitCx<'_>,
-    visible: &[&Piece<'_>],
-    key: &str,
-    is_plain_element: bool,
-) -> Result<(), EmitError> {
-    push_ident_key(cx, key);
-    cx.buf.push(": [");
-    let mut first = true;
-    for piece in visible.iter() {
-        if piece_event_key(piece, is_plain_element).as_deref() != Some(key) {
-            continue;
-        }
-        if !first {
-            cx.buf.push(", ");
-        }
-        first = false;
-        match piece {
-            Piece::On(on) => emit_on_value(cx, on)?,
-            Piece::ModelUpdate { source, .. } => super::model::emit_assignment(cx, source),
-            _ => {}
-        }
-    }
-    cx.buf.push("]");
-    Ok(())
-}
-
-pub(super) fn js_value<'a>(bind: &'a BindOp<'a>) -> Result<&'a JsExpr<'a>, EmitError> {
-    match bind.value {
-        Some(ExprRef::Js(js)) => Ok(js),
-        _ => Err(EmitError::Unsupported),
-    }
 }
